matting-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart
Pillow>=10.2.0
numpy
rembg>=2.0.50
onnxruntime>=1.17.0
